+++ config/config.exs
@@ -8,7 +8,8 @@
 use Mix.Config
 
 config :dealog_backoffice,
-  ecto_repos: [DealogBackoffice.Repo]
+  ecto_repos: [DealogBackoffice.Repo],
+  event_stores: [DealogBackoffice.EventStore]
 
 config :dealog_backoffice,
        DealogBackofficeWeb.Gettext,
@@ -30,6 +31,10 @@ config :logger, :console,
 
 # Use Jason for JSON parsing in Phoenix
 config :phoenix, :json_library, Jason
+
+# Configure event store
+config :commanded,
+  event_store_adapter: Commanded.EventStore.Adapters.EventStore
 
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.