+++ config/test.exs
@@ -18,6 +18,16 @@ config :dealog_backoffice, DealogBackoffice.Repo,
   hostname: System.get_env("POSTGRES_HOST") || "db",
   pool: Ecto.Adapters.SQL.Sandbox
 
+config :dealog_backoffice, DealogBackoffice.EventStore,
+  serializer: EventStore.JsonSerializer,
+  username: System.get_env("POSTGRES_USER") || "postgres",
+  password: System.get_env("POSTGRES_PASSWORD") || "postgres",
+  database:
+    System.get_env("POSTGRES_DB") ||
+      "dealog_backoffice_event_store_test#{System.get_env("MIX_TEST_PARTITION")}",
+  hostname: System.get_env("POSTGRES_HOST") || "db",
+  pool: Ecto.Adapters.SQL.Sandbox
+
 # We don't run a server during test. If one is required,
 # you can enable the server option below.
 config :dealog_backoffice, DealogBackofficeWeb.Endpoint,