+++ lib/dealog_backoffice/event_store.ex
@@ -1,0 +1,3 @@
+defmodule DealogBackoffice.EventStore do
+  use EventStore, otp_app: :dealog_backoffice
+end