+++ lib/dealog_backoffice/messages.ex
@@ -1,0 +1,5 @@
+defmodule DealogBackoffice.Messages do
+  @moduledoc """
+  The context for messages
+  """
+end